+++ src/lightning/pytorch/trainer/connectors/checkpoint_connector.py
@@ -401,7 +401,7 @@ class _CheckpointConnector:
         # restore the lr schedulers
         lr_schedulers = self._loaded_checkpoint["lr_schedulers"]
         for config, lrs_state in zip(self.trainer.lr_scheduler_configs, lr_schedulers):
-            config.scheduler.load_state_dict(lrs_state)
+            config["scheduler"].load_state_dict(lrs_state)
 
     def _restore_modules_and_callbacks(
         self, checkpoint_path: Optional[_PATH] = None, weights_only: Optional[bool] = None
@@ -469,7 +469,7 @@ class _CheckpointConnector:
             # dump lr schedulers
             lr_schedulers = []
             for config in trainer.lr_scheduler_configs:
-                lr_schedulers.append(config.scheduler.state_dict())
+                lr_schedulers.append(config["scheduler"].state_dict())
             checkpoint["lr_schedulers"] = lr_schedulers
 
             # precision plugin